add contacts

+++ vet/vetclient/src/types/schedule.cpp
@@ -217,7 +217,7 @@ void ScheduleList::add_shedule_item(Schedule &shed)
 	shedule_list.push_back(shed);
 }
 
-const Schedule& ScheduleList::at(size_t idx) const
+const Schedule& ScheduleList::at(int idx) const
 {
 	return shedule_list.at(idx);
 }
